requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
joblib==1.4.2
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
tabulate==0.9.0
tzdata==2025.1
urllib3==2.3.0
xgboost==3.0.0
scikit-learn==1.6.1
